# docs/swagger.yaml
basePath: /
definitions:
  github_com_rede_world-cup-quiniela_internal_middleware.ErrorDetail:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
  internal_api_handler.BulkOperationResultResponse:
    properties:
      failed:
        items:
          type: integer
        type: array
      succeeded:
        items:
          type: integer
        type: array
    type: object
...
  internal_api_handler.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/github_com_rede_world-cup-quiniela_internal_middleware.ErrorDetail'
    type: object
...
  internal_api_handler.GroupResponse:
    properties:
      created_at:
        type: string
      currency:
        type: string
      entry_fee:
        type: integer
      id:
        type: integer
      invite_code:
        type: string
      invite_code_expires_at:
        type: string
      max_members:
        type: integer
      name:
        type: string
      owner_user_id:
        type: integer
      prize_threshold:
        type: integer
      status:
        description: |-
          Status is system-managed: "active" when the group has ≥ 3 active members,
          "inactive" otherwise. Only active groups are eligible for payment processing
          and prize distribution.
        type: string
      updated_at:
        type: string
    type: object
...
  internal_api_handler.SnapshotEntryResponse:
    properties:
      prize_winner:
        type: boolean
      rank:
        type: integer
      total_points:
        type: integer
      user_id:
        type: integer
    type: object
  internal_api_handler.SnapshotResponse:
    properties:
      created_at:
        type: string
      entries:
        items:
          $ref: '#/definitions/internal_api_handler.SnapshotEntryResponse'
        type: array
      id:
        type: integer
      quiniela_id:
        type: integer
      taken_at:
        type: string
    type: object
...
  internal_api_handler.bulkGroupIDsRequest:
    properties:
      group_ids:
        items:
          type: integer
        type: array
    type: object
  internal_api_handler.bulkMemberIDsRequest:
    properties:
      membership_ids:
        items:
          type: integer
        type: array
    type: object
...
  internal_api_handler.transferOwnershipRequest:
    properties:
      new_owner_user_id:
        type: integer
    type: object
  internal_api_handler.updateGroupSettingsRequest:
    properties:
      entry_fee:
        type: integer
      max_members:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: REST API for the World Cup prediction game. Manage fixtures, submit
    score forecasts, and track leaderboards.
  title: World Cup Quiniela API
  version: "1.0"
paths:
...
  /api/v1/admin/groups/{id}:
    delete:
      description: Soft-deletes the quiniela group. All memberships and predictions
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "403":
          description: Caller is not an admin
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "404":
          description: Group not found
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "422":
          description: Invalid group ID
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a group
      tags:
      - admin-groups
...
  /api/v1/admin/groups/{id}/leaderboard/recalculate:
    post:
      description: Triggers an immediate leaderboard snapshot for the group.
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handler.SnapshotResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "403":
          description: Caller is not an admin
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "422":
          description: Invalid group ID
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Recalculate leaderboard
      tags:
      - admin-groups
  /api/v1/admin/groups/{id}/members/{membershipID}:
    delete:
      description: Sets the given membership to the 'left' status, effectively
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      - description: Membership ID to remove
        in: path
        name: membershipID
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "403":
          description: Caller is not an admin
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "404":
          description: Membership not found or already inactive
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "422":
          description: Invalid membership ID
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove a member
      tags:
      - admin-groups
  /api/v1/admin/groups/{id}/members/bulk-remove:
    post:
      consumes:
      - application/json
      description: Sets multiple memberships to 'left'. Returns 200 when all
      parameters:
      - description: Group ID - only memberships belonging to this group are removed
        in: path
        name: id
        required: true
        type: integer
      - description: Membership IDs to remove
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api_handler.bulkMemberIDsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handler.BulkOperationResultResponse'
        "207":
          description: Multi-Status
          schema:
            $ref: '#/definitions/internal_api_handler.BulkOperationResultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "403":
          description: Caller is not an admin
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "422":
          description: membership_ids is required or exceeds limit
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Bulk remove members
      tags:
      - admin-groups
...
  /api/v1/admin/groups/{id}/settings:
    patch:
      consumes:
      - application/json
      description: Changes the max_members cap and entry_fee of a group atomically.
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      - description: Settings to update (entry_fee required)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api_handler.updateGroupSettingsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handler.GroupResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "403":
          description: Caller is not an admin
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "404":
          description: Group not found
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "422":
          description: entry_fee is required
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update group settings
      tags:
      - admin-groups
  /api/v1/admin/groups/{id}/transfer-ownership:
    post:
      consumes:
      - application/json
      description: Assigns MembershipRoleCreateOwner to new_owner_user_id and demotes
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      - description: New owner user ID
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api_handler.transferOwnershipRequest'
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "403":
          description: Caller is not an admin
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "404":
          description: Group or new owner not found
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "422":
          description: new_owner_user_id is required
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Transfer group ownership
      tags:
      - admin-groups
  /api/v1/admin/groups/bulk-delete:
    post:
      consumes:
      - application/json
      description: Soft-deletes multiple quiniela groups. Returns 200 when all
      parameters:
      - description: Group IDs to delete
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api_handler.bulkGroupIDsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handler.BulkOperationResultResponse'
        "207":
          description: Multi-Status
          schema:
            $ref: '#/definitions/internal_api_handler.BulkOperationResultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "403":
          description: Caller is not an admin
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "422":
          description: group_ids is required
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Bulk delete groups
      tags:
      - admin-groups
...
securityDefinitions:
  BearerAuth:
    description: 'Clerk JWT token. Format: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"